+++ src/common/webviews/views/aurafx.ts
@@ -1,12 +1,9 @@
 import { openToolExternally } from '../common';
 
-const aurafxLink = 'https://aurafx.vercel.app?source=MythicScribe';
+const aurafxLink = 'https://aurafx.vercel.app';
 //const name = 'AuraFX by Sleepsweety';
 
 export function openAuraFXWebview(): void {
-    // if (!!getGenericConfig('allowExternalTools') === false) {
-    //     return externalToolWarning(aurafxLink);
-    // }
     openToolExternally(aurafxLink);
     return;
 }