+++ src/common/webviews/views/minecraftsounds.ts
@@ -1,0 +1,8 @@
+import { LinkParameters, openToolExternally } from '../common';
+
+const link = 'https://minecraftsounds.com';
+
+export function openMinecraftSoundsWebview(sound?: string, anchor?: LinkParameters): void {
+    openToolExternally(link, sound ? { search: sound } : undefined, anchor);
+    return;
+}