+++ src/web/datasets/ScribeMechanic.ts
@@ -7,6 +7,6 @@ export async function getDirectoryFiles(
     registry: AbstractScribeMechanicRegistry
 ): Promise<vscode.Uri[]> {
     return registry.files.map((file) =>
-        vscode.Uri.joinPath(ctx.extensionUri, 'data', registry.folder, file + '.json')
+        vscode.Uri.joinPath(ctx!.extensionUri, 'data', registry.folder, file + '.json')
     );
 }