+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Decorations for delay mechanics inside of metaskills, showing the accumulated delay since the start of the metaskill itself. Can be enabled via the configuration `MythicScribe.decorationOptions.delayTracking`
+- Decorations for sound mechanics, showing a prompt to play the sound on minecraftsounds.com
 
 ### Fixed
 - Mythic Node View's "Go To Edge" feature returning wrong range when used against the edge of a template